another major pupdate

+++ smash-ranks-android/app/src/main/java/com/garpr/android/sync/RankingsPollingSyncManagerImpl.kt
@@ -58,7 +58,7 @@ class RankingsPollingSyncManagerImpl(
             pollFrequency = PollFrequency.DAILY
         }
 
-        jobBuilder.trigger = Trigger.executionWindow(0, pollFrequency.timeInSeconds)
+        jobBuilder.trigger = Trigger.executionWindow(0, pollFrequency.timeInSeconds.toInt())
         jobDispatcher.mustSchedule(jobBuilder.build())
 
         mTimber.d(TAG, "sync has been enabled")
